feat: add access-control-expose-headers header to responses
For the new POST /topic/ method to work, we need to be able to read the topicId from the 'Location' header, which isn't exposed in CORS requests by default.  Adding this header will allow it to be read from javascript.  Frustratingly, this doesn't show up in a node test environment, because it doesn't have the same CORS restrictions